chore: packages/ai-rag setup
