+++ packages/ai-rag/eslint.config.mjs
@@ -1,0 +1,4 @@
+import { config } from "@repo/eslint-config/base";
+
+/** @type {import("eslint").Linter.Config} */
+export default config;
