+++ packages/ai-rag/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@repo/pinecone",
+  "name": "@repo/ai-rag",
   "version": "0.0.0",
   "private": true,
   "type": "module",
@@ -25,8 +25,9 @@
     "typescript": "^5.8.3"
   },
   "dependencies": {
-    "@pinecone-database/pinecone": "^6.0.0",
+    "@repo/pinecone": "workspace:*",
     "@repo/utils": "workspace:*",
+    "ai": "^4.3.15",
     "zod": "^3.24.4"
   }
 }
