+++ packages/pinecone/src/service.ts
@@ -1,0 +1,3 @@
+export async function upsertResources() {}
+
+export async function findRelevantContext() {}
